Revert cell status component
